moving model to public location for sharing
